Task #1418: Returning INT_MAX for maxTime iso LONG_MAX to make it work on 64 bit platforms also.

+++ trunk/MAC/APL/RTCCommon/include/APL/RTCCommon/Timestamp.h
@@ -123,7 +123,7 @@ namespace RTC
 	   */
 	  long sec()  const;
 	  long usec() const;
-	  static Timestamp maxTime() { return (Timestamp(LONG_MAX, 999999)); };
+	  static Timestamp maxTime() { return (Timestamp(INT_MAX, 999999)); };
 	  /*@}*/
 
 public:
